Coalesce StatementConfiguration, SqlStatements config classes
