+++ core/src/main/java/org/jdbi/v3/core/result/ResultProducers.java
@@ -18,7 +18,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.function.Supplier;
 
-import org.jdbi.v3.core.statement.StatementConfiguration;
+import org.jdbi.v3.core.statement.SqlStatements;
 import org.jdbi.v3.core.statement.StatementContext;
 
 /**
@@ -78,7 +78,7 @@ public class ResultProducers {
      */
     public static ResultProducer<ResultSetIterable> returningGeneratedKeys(String... generatedKeyColumnNames) {
         return (supplier, ctx) -> {
-            StatementConfiguration cfg = ctx.getConfig(StatementConfiguration.class);
+            SqlStatements cfg = ctx.getConfig(SqlStatements.class);
             cfg.setReturningGeneratedKeys(true);
 
             if (generatedKeyColumnNames.length > 0) {
